+++ src/openxr/include/signals_util.h
@@ -7,6 +7,10 @@ using namespace godot;
 
 static const char *SIGNAL_SESSION_BEGUN = "openxr_session_begun";
 static const char *SIGNAL_SESSION_ENDING = "openxr_session_ending";
+static const char *SIGNAL_SESSION_IDLE = "openxr_session_idle";
+static const char *SIGNAL_SESSION_SYNCHRONIZED = "openxr_session_synchronized";
+static const char *SIGNAL_SESSION_LOSS_PENDING = "openxr_session_loss_pending";
+static const char *SIGNAL_SESSION_EXITING = "openxr_session_exiting";
 static const char *SIGNAL_FOCUSED_STATE = "openxr_focused_state";
 static const char *SIGNAL_VISIBLE_STATE = "openxr_visible_state";
 static const char *SIGNAL_POSE_RECENTERED = "openxr_pose_recentered";
@@ -21,6 +25,10 @@ static void register_plugin_signals() {
 	// Register our signals
 	arvr_server->add_user_signal(SIGNAL_SESSION_BEGUN);
 	arvr_server->add_user_signal(SIGNAL_SESSION_ENDING);
+	arvr_server->add_user_signal(SIGNAL_SESSION_IDLE);
+	arvr_server->add_user_signal(SIGNAL_SESSION_SYNCHRONIZED);
+	arvr_server->add_user_signal(SIGNAL_SESSION_LOSS_PENDING);
+	arvr_server->add_user_signal(SIGNAL_SESSION_EXITING);
 	arvr_server->add_user_signal(SIGNAL_FOCUSED_STATE);
 	arvr_server->add_user_signal(SIGNAL_VISIBLE_STATE);
 	arvr_server->add_user_signal(SIGNAL_POSE_RECENTERED);